+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Use object `Configuration` for getting base module configuration
 - Add possibility to select mode(whitelist/blacklist) in ProxyFilter.php
     * The default option is blacklist
+- Allow call multiple ProcessFilter in one ProxyFilter module
 
 #### Fixed
 - Fixed the width of showed tagged idps in case the count of idps is equal to (x * 3) + 1